Refactor lithological description component
