+++ src/client/src/commons/form/profile/components/profileLayers/components/lithologicalDescriptionLayers/lithologicalDescriptionDisplay.js
@@ -1,0 +1,31 @@
+import React from "react";
+import { Stack, Typography } from "@mui/material";
+import { useTranslation } from "react-i18next";
+
+const LithologicalDescriptionDisplay = props => {
+  const { item } = props;
+  const { t, i18n } = useTranslation();
+
+  return (
+    <Stack direction="column" justifyContent="space-between">
+      <Typography variant="subtitle1">{item.fromDepth} m</Typography>
+      <Typography
+        sx={{
+          fontWeight: "bold",
+          overflow: "auto",
+          display: "-webkit-box",
+          WebkitLineClamp: "2",
+          WebkitBoxOrient: "vertical",
+        }}>
+        {item.description}
+      </Typography>
+      {item.id !== null && (
+        <Typography variant="subtitle2">
+          {t("qt_description")}: {item.qtDescription?.[i18n.language] ?? "-"}
+        </Typography>
+      )}
+      <Typography variant="subtitle1">{item.toDepth} m</Typography>
+    </Stack>
+  );
+};
+export default LithologicalDescriptionDisplay;
